Updates to get java SWIG compiling again, handled swig wrapping of boost types and removing of consts

+++ APL/CommandTypes.cpp
@@ -135,7 +135,7 @@ namespace apl
 
 	}
 
-	BinaryOutput::BinaryOutput(ControlCode aCode,boost::uint8_t aCount, boost::uint8_t aOnTime, boost::uint8_t aOffTime) :
+	BinaryOutput::BinaryOutput(ControlCode aCode,boost::uint8_t aCount, boost::uint32_t aOnTime, boost::uint32_t aOffTime) :
 	CommandRequest(CT_BINARY_OUTPUT),
 	mRawCode(aCode),
 	mCount(aCount),
